Simplify typing.

+++ SoundFontInfoLib/SoundFontInfoLib/Entity/Generator/Amount.hpp
@@ -35,10 +35,10 @@ public:
   int16_t signedAmount() const { return raw_.shAmount; }
   
   /// @returns low value of a range (0-255)
-  short low() const { return int(raw_.ranges[0]); }
-  
+  int low() const { return int(raw_.ranges[0]); }
+
   /// @returns high value of a range (0-255)
-  short high() const { return int(raw_.ranges[1]); }
+  int high() const { return int(raw_.ranges[1]); }
   
 private:
